move examples out of operators module
