+++ lib/operators.js
@@ -134,50 +134,11 @@ ternaryR f op1 op2 p = (g <$> p <*> op1 <*> recur <*> op2 <*> recur) <|> p -- TO
     g a1 _ a2 _ a3 = f a1 a2 a3
 */
 
-var mult = chainR2(C.app(function(o) {return function(l, r) {return ['(', l, ' ', o, ' ', r, ')'].join('');};},
-                         C.position.oneOf('*/')),
-                   C.position.oneOf('abcdefg')),
-    add = chainR2(C.app(function(o) {return function(l, r) {return ['(', l, ' ', o, ' ', r, ')'].join('');};},
-                        C.position.oneOf('+-')),
-                  mult);
-
-var m2 = chainL2(C.app(function(o) {return function(l, r) {return ['(', l, ' ', o, ' ', r, ')'].join('');};},
-                      C.position.oneOf('*/')),
-                C.position.oneOf('abcdefg')),
-	a2 = chainL2(C.app(function(o) {return function(l, r) {return ['(', l, ' ', o, ' ', r, ')'].join('');};},
-                      C.position.oneOf('+-')),
-                m2);
-
-function node(op) {
-    return function(l, r) {
-        return {'l': l, 'r': r, 'op': op};
-    };
-}
-
-function dump(n) {
-    if (!n.l || !n.r) { console.log('oops -- ' + JSON.stringify(n)); }
-    var l = (typeof n.l === 'string') ? n.l : dump(n.l),
-        r = (typeof n.r === 'string') ? n.r : dump(n.r);
-    return ['(', l, ' ', n.op, ' ', r, ')'].join('');
-}
-
-var letter = C.position.satisfy(function(c) {return c >= 'a' && c <= 'z';}),
-    word = C.fmap(function(x) {return x.join('');}, C.many1(letter)),
-    a = chainL(C.app(node, C.position.string('**')), word),
-    b = chainR2(C.app(node, C.position.literal('^')), a),
-    c = chainL2(C.app(node, C.position.oneOf('+-')), b);
-
-var q = c.parse('abc**def**ghi^jkl^mno+pqr+stu', [1,1]);
-
 module.exports = {
     'chainL': chainL,
+    'chainL2': chainL2,
     'chainR': chainR,
-    
-    'add': add,
-    'mult': mult,
-    'c': c,
-    'eg1': add.parse('a+b*c*d/e+f-g', [1,1]),
-    'eg2': a2.parse('a+b*c*d/e+f-g', [1,1]),
-    'eg3': q,
-    'eg3s': dump(q.value.result)
+    'chainR2': chainR2,
+    'prefix': prefix,
+    'postfix': postfix,
 };
